:recycle: Changed system prompt to fit vegan scanner app, added logging and refactored load model output the svelte way
