+++ frontend/src/routes/food-scan/+page.ts
